\name{ec2test-package}
\alias{ec2test-package}
\alias{ec2test}
\docType{package}
\title{
\packageTitle{ec2test}
}
\description{
\packageDescription{ec2test}
}
\details{

The DESCRIPTION file:
\packageDESCRIPTION{ec2test}
\packageIndices{ec2test}
~~ An overview of how to use the package, including the most important functions ~~
}
\author{
\packageAuthor{ec2test}

Maintainer: \packageMaintainer{ec2test}
}
\references{
~~ Literature or other references for background information ~~
}
~~ Optionally other standard keywords, one per line, from file KEYWORDS in the R ~~
~~ documentation directory ~~
\keyword{ package }
\seealso{
~~ Optional links to other man pages, e.g. ~~
~~ \code{\link[<pkg>:<pkg>-package]{<pkg>}} ~~
}
\examples{
~~ simple examples of the most important functions ~~
}
